+++ Tim17_Back/src/main/java/com/ib/Tim17_Back/validations/UserRequestValidation.java
@@ -47,11 +47,11 @@ public class UserRequestValidation {
         String token = headers.get("x-auth-token");
         String refreshToken = headers.get("refreshtoken");
         if(isTokenExpired(token)){
-            Integer id = jwtTokenUtil.getId(refreshToken);
+            Long id = jwtTokenUtil.getId(refreshToken);
             return givenId.intValue() == id;
 
         }
-        Integer id = jwtTokenUtil.getId(token);
+        Long id = jwtTokenUtil.getId(token);
         return givenId.intValue() == id;
     }
